requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pdf2image>=1.16.0
PyPDF2>=3.0.0

# AI and ML
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Microsoft Graph API
msal>=1.24.0
msgraph-core>=1.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0



